pyytsync/requirements.txt
django-bootstrap-v5
django-bootstrap-icons
isodate
google-api-python-client
